wrap PyArg_ParseTuple and Py_BuildValue
Automatically deduce the format mapping to ensure right one is used in
order to not have type errors, and to detect if the actual C
type (e.g. with DWORD-ish typedefs) actually uses the right python
type. For example, `HANDLE` may map to `unsigned long` on arch A but
to `unsigned int` on arch B. This is now handled.

If compilation fails for a missing tuple_char_for_type specialization,
the type-to-character mapping needs to be added in the (currently)
short list of known types in python_wrapper.hpp
(`tuple_char_for_type` specializations). This should only be needed
once for new architectures, especially when switching 32 to 64 bit or
the other way around.