manually check types in StormLib API
this sadly is a manual process since C++ is full of implicit
conversions and StormLib does not use hard typedefs. These types are
in sync with Stormlib 9.22. Since the API should rarely change, I hope
this is a one time thing, not something one has to repeat every few
releases.